lib/wrappers/python/requirements.txt
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
mediapipe==0.9.3.0
pyserial==3.5
types-pyserial==3.5.*
mypy==1.10.*